select_limit was not checked and honoured when changing screen mode (#6)
select_limit was not checked and honoured

This caused an issue with ChangeResolutionOptions forgetting that the user should not be able to check any more of the resolution option boxes
I don't believe the inverse needs to be implemented, i.e. to enable all checkboxes if the limit isn't reached.