Disable text boxes when disabling check boxes
Disables/Enables text boxes which are not associated with a selected resolution. This mirrors the behaviour of the checkboxes (and now means the custom resolutions behave the same as the selectable ones).

Any values in these unselected boxes are not saved anyway, so you only need to edit them if you have/are going to select them. This also makes it clearer to the user that the limit has been reached.