Resize window on creation for High DPI
Another quick fox to help higher DPI users.

Running Fit(); will resize the window to fit all of the contents. This is needed for users with a DPI set above "100%", otherwise the window is too small to see all of the options.

There is a check for 600p, that is mainly for forward compatability with the unofficial branch. Running Fit() may adjust the dimensions of the window slightly from what is configured above, in the case of the unofficial branch, this would make the window too tall at 600p. Hence the if statement. For any higher resolution, this growth is irrelevant, and the benefit of the wxWidget calculation, is more evenly spaced checkboxes etc.

There is some additional work for higher DPI settings, but this would require a resonable amount of work. Changing values from pixels to points, and (potentially from what I read) not relying on static boxes etc.

I will attach an image of the new results, you can see that some of the elements (text boxes and radiobutton labels) are not ideal, the textboxes can be improved by expanding the window's width, the labels cannot be improved by resizing the window. The Launcher window needs a bit of work, but is functional at 200% (the picture is 150%).