failsafe is no longer added to INGAME_RES list automatically
MAX_RESOLUTIONS made global and set to 6 - this represents the size of the `switching_vidmodes` array in KFX. Code refactored to use this constant; which indicates the number of "in-game resolutions" that the user can set in the Launcher.

Failsafe res will not be factored in to the list of in-game resolutions UNLESS the user fails to select any resolutions in the settings window of the launcher, then the failsafe will be added as the only entry for INGAME_RES. On a subsequent load of the settings window, this added "failsafe" resolution will be ticked in the In-game resolution section.